% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovacircR.test.R
\name{anovacircR.test}
\alias{anovacircR.test}
\title{anovacircR.test: two-dimensional analysis of variance using complex data
an extension of the logic of the t-squared circ statistic
this is the one-way, repeated measures implementation of the test
the expected input is a long-format data frame with the following columns:
simdata: the data stored as complex values
grouplabels: condition labels indicating the level of the independent variable each data point corresponds to
participant: participant ID number
see Baker (2021) for further details}
\usage{
anovacircR.test(data)
}
\description{
anovacircR.test: two-dimensional analysis of variance using complex data
an extension of the logic of the t-squared circ statistic
this is the one-way, repeated measures implementation of the test
the expected input is a long-format data frame with the following columns:
simdata: the data stored as complex values
grouplabels: condition labels indicating the level of the independent variable each data point corresponds to
participant: participant ID number
see Baker (2021) for further details
}
