% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.test.R
\name{CI.test}
\alias{CI.test}
\title{function to calculate the condition index of bivariate data
compare for significance with an expected distribution function}
\usage{
CI.test(data)
}
\description{
function to calculate the condition index of bivariate data
compare for significance with an expected distribution function
}
