% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustercorrect.R
\name{clustercorrect}
\alias{clustercorrect}
\title{clustercorrect: implements non-parametric cluster correction method described by Maris & Oostenveld (2007) J Neurosci Methods, 164: 177-190, doi: 10.1016/j.jneumeth.2007.03.024
this version can deal with complex data using the T-squared or T-squared-circ statistics}
\usage{
clustercorrect(
  datax,
  datay = NULL,
  adjacencymatrix = NULL,
  testtype = 1,
  paired = TRUE,
  clustformthresh = 0.05,
  clustthresh = 0.05,
  nresamples = 10000
)
}
\description{
The input datax should be an N (participants) x m (sensors/timepoints/locations) matrix
The optional input datay can either be a matrix of the same dimensions, or N2 x m (for unpaired designs with unbalanced samples), or a single value to compare to datax
}
\details{
Legal values of the testtype input are 1: t-test, 2: Hotelling's T-squared, 3: Victor & Mast's T-squared-circ
Complex values are required for options 2 & 3. If complex values are passed in for test 1 (t-test), they will be converted to amplitudes

if the measures are from adjacent time points, an adjacency matrix will be automatically constructed
if the measures are from different spatial locations (e.g. an electrode montage) you should calculate an adjacency matrix yourself
this should be an m x m matrix (where m is the number of sensors), where values are zero for non-adjacent sensor pairs, and 1 for adjacent pairs

outputs an object containing the indices of all significant clusters
}
