% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getel.R
\name{getel}
\alias{getel}
\title{helper function that calculates an ellipse bounding some data points
the outline of the ellipse is returned}
\usage{
getel(compdata)
}
\description{
helper function that calculates an ellipse bounding some data points
the outline of the ellipse is returned
}
