% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwisemahal.R
\name{pairwisemahal}
\alias{pairwisemahal}
\title{pairwisemahal: calculates the pairwise (group to group) Mahalanobis distance between group means}
\usage{
pairwisemahal(data, grouping = NULL)
}
\description{
inputs:
data - can be either a vector of N complex numbers, a matrix of Nxm complex numbers, or an Nx2 matrix of real numbers
if a vector of complex numbers or matrix of real numbers, the 'grouping' variable is also required
if a matrix of complex numbers is supplied, the columns are interpreted as the groups
grouping - a variable containing group membership information for the data
}
\details{
output:
an mxm data frame of Mahalanobis distances between the group means
note that the Mahalanobis distance (D) is returned, not D^2
the row and column names of the data frame are the condition names from the grouping input variable (or are otherwise numeric)

this function is part of the FourierStats package: https://github.com/bakerdh/FourierStats
}
