% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsq1.test.R
\name{tsq1.test}
\alias{tsq1.test}
\title{function to calculate a one-sample t-squared test after Hotelling (1931)
for two-sample tests, use the Hotelling package}
\usage{
tsq1.test(data, mu = NULL)
}
\description{
function to calculate a one-sample t-squared test after Hotelling (1931)
for two-sample tests, use the Hotelling package
}
