% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsqc.test.R
\name{tsqc.test}
\alias{tsqc.test}
\title{function to calculate the t-squared circ statistic from Victor & Mast (1991)
the inputs must be Nx2 or 2xN matrices of numbers
for a one-sample test, mu is an optional vector to which the data are compared}
\usage{
tsqc.test(x, y = NULL, paired = FALSE, mu = NULL)
}
\description{
function to calculate the t-squared circ statistic from Victor & Mast (1991)
the inputs must be Nx2 or 2xN matrices of numbers
for a one-sample test, mu is an optional vector to which the data are compared
}
