% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsqh.test.R
\name{tsqh.test}
\alias{tsqh.test}
\title{tsqh.test: function to calculate a different variants of the T-squared test of Hotelling (1931)
the inputs must be Nx2 or 2xN matrices of numbers, or vectors of complex numbers
for a one-sample test, mu is an optional vector to which the data are compared
for a two-sample test, x and y are the data from the two conditions
if paired=TRUE, the inputs (x and y) must be the same size}
\usage{
tsqh.test(x, y = NULL, paired = FALSE, mu = NULL)
}
\description{
tsqh.test: function to calculate a different variants of the T-squared test of Hotelling (1931)
the inputs must be Nx2 or 2xN matrices of numbers, or vectors of complex numbers
for a one-sample test, mu is an optional vector to which the data are compared
for a two-sample test, x and y are the data from the two conditions
if paired=TRUE, the inputs (x and y) must be the same size
}
