% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysecplx.R
\name{analysecplx}
\alias{analysecplx}
\title{analysecplx: wrapper function that decides which statistical test to run for a given data set
the expected input is a data frame in long format:
one column should contain complex valued data, or two columns headed x and y contain the real and imaginary parts
if a column is headed 'group' (or any word beginning with a g) is included, this will be interpreted as level data for the independent variable
if a column is headed 'participant' (or any word beginning with a p) is included, this will be interpreted as a random factor}
\usage{
analysecplx(input)
}
\description{
the function first runs the condition index test for each level of the IV
if the condition index test is non-significant for all levels, a T-squared-circ test is run for one- and two-sample designs, or an ANOVA-squared-circ test otherwise
if the condition index test is significant for any level, a T-squared or MANOVA test is run instead
the paired-samples Mahalanobis effect size statistic is also calculated
}
