% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovacirc.test.R
\name{anovacirc.test}
\alias{anovacirc.test}
\title{analysis of variance using complex data
an extension of the logic of the t-squared circ statistic}
\usage{
anovacirc.test(data)
}
\description{
analysis of variance using complex data
an extension of the logic of the t-squared circ statistic
}
