% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovacircR.test.R
\name{anovacircR.test}
\alias{anovacircR.test}
\title{analysis of variance using complex data
an extension of the logic of the t-squared circ statistic
this version for repeated measures designs}
\usage{
anovacircR.test(data)
}
\description{
analysis of variance using complex data
an extension of the logic of the t-squared circ statistic
this version for repeated measures designs
}
