% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.test.R
\name{CI.test}
\alias{CI.test}
\title{CI.test: function to calculate the condition index of bivariate data
Inputs:
data is an N x 2 matrix of bivariate (x,y) observations, or a vector of N complex values
alpha is the criterion for statistical significance, set to a default of 0.05
the condition index is square root of the the ratio of eigenvalues (eigenvector lengths), calculated longest/shortest
the index is then compared for significance with an expected distribution function
significant tests (p < alpha) violate the assumptions of the T-squared-circ and ANOVA-squared-circ tests
see Baker (2021, http://arxiv.org/abs/2101.04408) for further details}
\usage{
CI.test(data, alpha = 0.05)
}
\description{
CI.test: function to calculate the condition index of bivariate data
Inputs:
data is an N x 2 matrix of bivariate (x,y) observations, or a vector of N complex values
alpha is the criterion for statistical significance, set to a default of 0.05
the condition index is square root of the the ratio of eigenvalues (eigenvector lengths), calculated longest/shortest
the index is then compared for significance with an expected distribution function
significant tests (p < alpha) violate the assumptions of the T-squared-circ and ANOVA-squared-circ tests
see Baker (2021, http://arxiv.org/abs/2101.04408) for further details
}
