% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getel.R
\name{getel}
\alias{getel}
\title{getel: helper function that calculates an ellipse bounding some data points
the outline of the ellipse is returned, sampled at 200 points
the input should be an Nx2 matrix of observations}
\usage{
getel(compdata)
}
\description{
getel: helper function that calculates an ellipse bounding some data points
the outline of the ellipse is returned, sampled at 200 points
the input should be an Nx2 matrix of observations
}
